Flesh out the logging implementation.
Add PLOG and ensure we have CHECK_STREQ and CHECK_STRNE to go with
the existing non-functional DCHECK_STREQ and DCHECK_STRNE.

This also gives us two different logging implementations, one for
the host that just uses std::cerr, and one for Android that uses
the Android log.

Also bring in the StringPrintf family.

Change-Id: I8e190c2c58f26b22ee76a2a87d447df6eb0fa73b