Fixes for register promotion; enable fast path
Had an off-by-one error in the range argument loading (only mattered when
register promotion is enabled - for range args we do a memory copy and
must ensure that all promoted values are flushed to memory before the
copy.  Also reworked and removed some asserts for sitations that are
legal now that we're promoting.  Enabled fast path code to exercise it
a bit.

Change-Id: Id1acb3dad01d5d1077661150e98e51fd4243b6f1