Test conversion of floats and doubles to strings.
There was some confusion about what Method::HasCode meant, and we weren't
quite ready to compile all methods _and_ be able to invoke them. We were
also missing a couple of native methods in Throwable that we need if we've
compiled all Throwable's code (because the next time we ask ClassLinker to
do anything, it'll try to throw NoClassDefFoundException from one of the
ClassLoaders, and that will try to run a Throwable constructor, which will
end up trying to call these native methods).

Change-Id: If4783f3c866aaa72413d7b7810ef2541d418ae33