Added more checking upon resolution and bytecode rewriting to verifier.
Access checks are performed when methods and fields are resolved. Also,
erroring bytecodes are now overwritten in the memory mapped dex file.
To do this, the code sets the memory mapped dex file as writable before
verification and set it back to read only after verification is done.
The overwritting occurs only in memory and the original dex file remains
unchanged.

Change-Id: I054394fb57e83d1ac5b6f200ab993d70cd9f55e6