Fix the 051-thread test.
This changes Thread::Create to clean up the requested stack size before
passing it to pthread_attr_setstacksize.

I modified the test to work around the inherent raciness in the test, which
currently causes us trouble because with all the debug logging, the thread
that initializes System.out takes longer than 1ms, so the first 40 or so
threads all get woken in non-numeric order.

Change-Id: I8047bf00330aaf4210dcb53a4437b1fe0562179e