Fix field access.
There was a bug in the implementation where the native methods weren't
necessarily getting a null receiver but art::Object insists that you only
get/set static fields on a null receiver.

Also, these tests assume that fields are in the setAccessible(false) state
by default. In our system, they're not.

Change-Id: I935614fec013be465fc0bc30d3c717f12476e09b