Warn if a GC takes > 100ms, even if -verbose:gc isn't on.
I haven't actually seen this in practice, but it does mean we can
rule GC out if we notice a pause and there's no GC logging. This might
also be useful if we change the launcher so apps aren't all launched
with -verbose:gc, as seems likely to happen eventually.

Change-Id: I06eba14740307eb75f70a4fab1f81443d18ebc3a