Improve exceptions thrown during class loading.
We've always gone to a lot of trouble to have good detail messages when
something goes wrong during class loading, but none of those exceptions
would ever make it to the calling code. This adds missing exception
chaining, and also adds a test for the case where a subclass is missing
its superclass; before we'd have reported that the subclass was missing,
but now we make it clear that the real error was an inability to find
the superclass.

Change-Id: I07ebc011ccdaed16be82bf08b323393e1d790989