Fix a thread state misunderstanding.
Note that this means that we now give the right answer for Object.wait(long)
but the wrong answer for Thread.sleep(long). It's not obvious how to easily
fix the latter. It seems like the easiest fix will be to check the class of
the object that's being waited on, and using TS_SLEEPING if it's a ThreadLock;
we already do similar in the stack dumping.

Change-Id: I83c0c54af1061e1678cefe7b07389dd10c2ea3d5