Suppress compilation of malformed dexfiles to prevent compiler segfault.
Generic failures detected at compile time will not be fixed at runtime,
so classes with generic failures are now added to a rejected classes
set and not compiled at all. This prevents the compiler from
segfaulting on certain occasions when registers and indexes are out of
range, etc. There is one remaining segfault in the vm-tests now.

Change-Id: Id67c12fd13f3e993a6a16a8625620daa0ea496cb