Verifier can handle an unresolved method or class.
Added a new dex verifier entry point that takes in a ClassDef index and
context instead of a Class pointer. This allows basic structural
verification to still be run on classes that FindClass fails to find. If
the basic structural classes fail, then the compilation for the class is
skipped. Before, an unresolvable malformed class would cause the
compiler to segfault.

Change-Id: Ia02c77560bf274281c4e22250983dc96511501eb