Unify branch flags, pretty printer for OpKind.
LIR operand 0 is always an offset for a branch. This is clear in
conditional branches that are binary and have the 2nd operand as the
condition codes of the branch. This changes unconditional branches to be
unary and therefore more intention revealing that the 1st operand will
be used by the assembler to hold an offset.

A << operator for OpKind allows easy pretty printing.

Change-Id: I933b8e0bf43f5be3eff13f93c3fc1539ae526840