Fix getter/setter special case codegen
The special-purpose code generators for simple methods that get
or set and instance field and then return require that no throws
are possible.  The previous code incorrectly relied on the first
argument being a "this" pointer, and thus previously null-checked.
This did not take into account the possibility of a static method
which happened to pass an object referece as it's first argument.

The fix is to avoid making any assumptions, but rather rely
solely on the results of the null-check elimination pass which
will correctly recoginize the "this" case.

Change-Id: Icf001a10a19234cf3f4d87cf1baede93fdf0360c