Do not reset SirtRef pointing to a live object
ClassLinker::FindClass reset the klass SirtRef to point to the
existing Class value. However, the old value was still referenced by
the ObjectLock. In the rare case of a two thread resolving the same
class at the same time and a garbage collection happening, the
ObjectLock would point to freed memory.

Change-Id: I93dbbfe3e5d7a8922464242270ac90c71a125e47