Improve our check for whether a thread is still starting up.
It turns out that there was a race in Thread::Create that could confuse our
suspend-count sanity checking. Tested by manually inserting a sleep(3) there.
I could have added an extra field to Thread, but since we only need to check
this once if we do a GC while a thread is still starting up, a simple string
comparison will be fine.

  thread_list.cc:79] Thread[10,tid=7447,VmWait,Thread*=0x00d23060,peer=0x60f1b350,"<native thread without managed peer>"] suspend count already zero

I've also added some logging so we'll be nagged to remove a work-around for
a bionic bug when we're in a tree where the bionic bug is fixed.

Change-Id: I78ac3c58245c1ecff3e86dcf297b94ae0085f7a2