Improve signal catching and native stack dumps.
This works around the problem with abort(3) (and friends) on glibc,
and improves our ability to catch SIGSEGVs from the main thread, and
slightly improves our behavior if multiple threads are dying at once.
I think any remaining flakiness is only in this last case, and I'm
not sure whether we can actually improve that any further.

Change-Id: I28a5347fc9c983766ea8802ae95fba36f96dbdc5