"InitGoogle"-style argv stashing.
This lets us give the command line in crash dumps when dex2oat dies in the
continuous build on the Mac. I've also taken the opportunity to use the
basename of argv[0] as the default log tag, so dex2oat will now show up in
logcat as "dex2oat" instead of "art" (and we can probably stop manually
prefixing dex2oat log output).

Also stash pthread_self() so we can _correctly_ report "handle=" in the
SIGQUIT output.

Change-Id: Ia8249cd19bab5b816cb94a531a65becdfacaa98b