Verifier treats unresolved return types as a soft error.
It's ok for the type of a returned register to be an unresolved class,
even if the return type derived from the descriptor is resolved, when
the descriptor type is a super and the register type is a sub. It's
also ok for both types to be resolved, but it's not ok for the derived
return type to be unresolved while the register's type is resolved.

Change-Id: If0ac5ba33ccef61e754382b7d9918deac7f05448