More Quick compiler restructuring
Yet another round of compiler restructuring to remove references to
MIR and BasicBlock structures in the lower levels of code generation.
Eliminating these will make it easer to share code with the
LLVM-IR -> LIR lowering pass.

This time I'm focusing on the Invokes and special-cased inlined
intrinsics.  Note that this CL includes a somewhat subtle difference
in handling MOVE_RETURNs following an INVOKE.  Previously, we fused
INVOKEs and MOVE_RETURNs for inlined intrinsics.  To simplify the
world, we're now fusing all INVOKE/MOVE_RETURN pairs (though it shouldn't
impact the codegen non-inlined invokes.

Changes in latest patch set: Two of the old intrinsic generators
did not fuse the following MOVE_RESULT, so the results were dropped
on the floor.  Also, allowed match of MOVE_RESULT_OBJECT (which
previously had not been matched because no inline intrisic used it
as a return value).

Change-Id: I93fec0cd557398ad7b04bdcc0393f27d3644913d