Fix string.compareTo regression
The recent restructuring that fused invokes and their subsequent
move-results caused a performance regression in the compareto
microbenchmarks.  The restructuring CL was being a little over conservative
and disabled detection of special inlined functions when encountering
one who's subsequent move-result was missing.  The string.compareto
microbenchmarks do a bunch of compareto's and ignore the result.

It's perfectly legal to ignore the return value, and the generaated
code handles this case just fine by defaulting back to the special
return registers if not move-result is found.  Remove the check.

Change-Id: Idca9d10404ced4d6b11be6ea00d62ed429345e85