Mod-union table implementation: reference caching
Implementation which uses a map of card pointers to reference array to keep track of image space references to the alloc space.

Speed/memory usage improvements over bitmap implementation. Approximated memory usage is ~4 * number_references compared to the previous 128k due a bitmap spanning the zygote space.

For the system server, memory usage is approximately 100k due to ~20000 references to the alloc space.

Performance increases since we no longer scan objects in the imagespace when we mark the references of the mod union table.

Change-Id: I449b8b3edeba712a535f6a3b14d81743bcd6f5a0