Ignore the redundant InferredRegCategoryMap.
The verifier may verify the same class more than once.
Thus, it is possible for us to "Put" the same key-value
pair to the "SafeMap" more than once.  In this case,
just ignore the redundant key-value pair.

NOTE: Since soft verification failure is not considered
as erroneous, we may invoke VerifyClass again while we
are verifying the subclass, which will verify its super
class if it is not successfully verified or in erroneous
state.

Change-Id: Ifbd817b6e00c9b452ebbf1140750025ff261ef1e