Quick compiler - NOT_LONG and verification fixes
We continue to have issues when the verifier rewrites Dalvik
byte codes.  This is a workaround to fix the GBC -> LIR
conversion, but I suspect I'm going to have to add a dead code
elimination pass to get things working properly for llvm.  The
basic problem is that if the re-written instruction defined a
Dalvik virtual register, then we lose the definition.  The
MIR to GBC converter doesn't handle uses without definitions
very well.

Change-Id: Ia235c299b686ce080755f9bb4c05c719c16c0d77