Remove a lock from ScopedThreadStateChange.
Remove a lock used while inspecting the current thread state in
ScopedThreadStateChange. It is safe to do so as the code just checks the
state and the suspend count checks occur in the transition logic (with
the appropriate locks held).

Gives small but noticeable performance improvement to the JniField
microbenchmark.

Change-Id: Iadda8ae0b30355dc3d7e4daeed760c41b3fecfc2