Port 138b7a2e54923d99ed817f9a890c6fcd7c23c335 to art
Disable ADDR_NO_RANDOMIZE for 3.4 or greater kernels

Early kernels had memory fragmentation bugs on ARM when
ASLR was enabled. The ADDR_NO_RANDOMIZE flag was added
to work around this bug until the appropriate kernel fix
could be made.

Linux kernel 3.4 seems to have this fix, making this workaround
unnecessary.  Remove the workaround if we detect we're running
an up-to-date kernel.

(I believe the kernel patch is also in 3.3, but haven't explicitly
tested it).

Bug: 5817320
Change-Id: Ife1e511355c62c16fc90ba22811face7b573b312