Split dex_pc<->native_pc mapping table
First of 2 related CLs - this one moves to a split table, a future
CL will compress the table to save image space.

The current MappingTable mixes two flavors of entries: dex->pc and
pc(return address)->dex.  The problem is that we can have two
entries with the same native pc address but two different dex pcs.
The reason is that when we go from native pc to dex, the native
pc is actually the return address from the previous call, whereas
when we go from dex to native pc, the mapping refers to the first
native instruction of the sequence.

Previously, the first entry in the mapping table was the number
of subsequent entries.  That will remain the same, but now
the second entry will be a count of the number of entries in
the pc->dex section of the table.  The difference between the
two counts gives us the number of entries in the following dex->pc
section.

Change-Id: Ibadb96cb50bf1f93e079dff3832130b9f9782723